feat: add system rpc

+++ server/system/rpc/model/authority.go
@@ -6,9 +6,9 @@ import (
 
 type Authority struct {
 	gorm.Model
-	AuthorityId   uint            `json:"authorityId" gorm:"not null;unique;primary_key;comment:role id for access control;size:90"`
+	AuthorityId   uint64          `json:"authorityId" gorm:"not null;unique;primary_key;comment:role id for access control;size:90"`
 	AuthorityName string          `json:"authorityName" gorm:"comment:role name"`
-	ParentId      uint            `json:"parentId" gorm:"comment:parent id"`
+	ParentId      uint64          `json:"parentId" gorm:"comment:parent id"`
 	AuthorityIds  []*Authority    `json:"AuthorityIds" gorm:"many2many:authority_ids;"`
 	Children      []Authority     `json:"children" gorm:"-"`
 	Menu          []AuthorityMenu `json:"menus" gorm:"many2many:authority_menus;"`
